+++ NEWS.md
@@ -1,4 +1,4 @@
-# metafor 4.1-6 (2023-04-24)
+# metafor 4.1-7 (2023-04-25)
 
 - also allow two or three values for `xlab` in the various `forest()` functions (for adding labels at the ends of the x-axis limits)
 
@@ -7,6 +7,8 @@
 - added `shade` and `colshade` arguments to the various `forest()` functions
 
 - added `link` argument to `rma.glmm()`
+
+- `rma.glmm(measure="OR", ..., model="CM.EL", method="ML")` now treats tau^2 values below 1e-04 effectively as zero before computing the standard errors of the fixed effects; this helps to avoid numerical problems in approximating the Hessian
 
 - for measure `SMCC`, can specify d-values, t-test statistics, and/or p-values via arguments `di`, `ti`, and `pi`